+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Unreleased
 - Introduced `symbolize::Input` enum and made it part of symbolization APIs to
   distinguish between and support different input types
   - Added support for ELF symbolization using file offsets instead of addresses
+- Added `symbolize::Source::Apk` variant
 
 
 0.2.0-alpha.7